if CHROME_ORIGIN is LOCAL_SOURCE, don't remove widevine patches

+++ chipset-rtd1619b/chromeos-base/chromeos-chrome/chromeos-chrome.bashrc
@@ -2,7 +2,13 @@ cros_pre_src_prepare_remove_internal() {
   if [ -d ${CHROME_ROOT}/src/google_apis/internal ]; then
     mv ${CHROME_ROOT}/src/google_apis/internal ${CHROME_ROOT}/src/google_apis/internal.bak
   fi
-  rm -rf ${CHROME_ROOT}/src/third_party/widevine/cdm/chromeos/*
+
+  if [ ${CHROME_ORIGIN} = "LOCAL_SOURCE" ]; then
+     echo "don't remove widevine dir"
+  else
+     echo "remove widevine dir"
+     rm -rf ${CHROME_ROOT}/src/third_party/widevine/cdm/chromeos/*
+  fi
 }
 
 PATCHES=(
